Added base class for OpenGL painting
OpenGlWidget is a simple base class that
implement the logic for painting on a widget.
After the creation you can call the function
container() to obtain a QWidget to use inside
the ui.
The subclass must reimplement initialize() and
draw()
With setAnimate(true) you can force the widget
to repaint as fast as possible